+++ src/swarmpal/toolboxes/fac/__init__.py
@@ -3,6 +3,6 @@ from swarmpal.toolboxes.fac.fac_algorithms import (
     fac_single_sat,
     fac_single_sat_algo,
 )
-from swarmpal.toolboxes.fac.fac_processor import FAC_singlesat
+from swarmpal.toolboxes.fac.processes import FAC_singlesat
 
 __all__ = ("fac_single_sat", "fac_single_sat_algo", "FacInputs", "FAC_singlesat")
