+++ src/swarmpal/toolboxes/secs/__init__.py
@@ -1,6 +1,8 @@
+from swarmpal.toolboxes.secs.processes import DSECS_Process
 from swarmpal.toolboxes.secs.secs_inputs import SecsInputs, SecsInputSingleSat
 
 __all__ = (
+    "DSECS_Process",
     "SecsInputSingleSat",
     "SecsInputs",
 )
